move storage into pkg
