+++ cmd/schema.go
@@ -55,5 +55,5 @@ var schemaCmd = &cobra.Command{
 func init() {
 	generateCmd.AddCommand(schemaCmd)
 	schemaCmd.Flags().StringSliceVarP(&schemaFiles, "schemas", "s", schemaFiles, "list of schema files to generate code from")
-	schemaCmd.Flags().StringVarP(&outputDirectory, "out", "o", "./storage/sqlite/schema", "directory to output generated code to")
+	schemaCmd.Flags().StringVarP(&outputDirectory, "out", "o", "./pkg/storage/sqlite/schema", "directory to output generated code to")
 }
