+++ pkg/storage/generate.go
