+++ pkg/storage/sqlite/schema/table/indexers.go
